pyproject.toml
[project]
name = "raghav"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.0.20",
    "fastapi>=0.116.1",
    "google-genai>=1.32.0",
    "langchain-google-genai>=2.1.10",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
